+++ c/helpers/stringlist.c
@@ -138,7 +138,7 @@ void StringList_Remove(StringList* stringList, int removeIndex)
 {
     if (stringList != NULL && stringList->strings != NULL)
     {
-        if (removeIndex >= 0 && removeIndex < stringList->strings_count)
+        if (removeIndex >= 0 && (size_t)removeIndex < stringList->strings_count)
         {
             const char* stringToRemove = stringList->strings[removeIndex];
             for (size_t stringIndex = removeIndex; stringIndex < stringList->allocated_count - 1; stringIndex++)
